Prevent dynamic changing of vterm buffer name
When an user customizes vterm-buffer-name-string to be non-nil, vterm
buffer names can change dynamically, depending, for example, on the
working directory of the shell. This causes problems in eval-in-repl,
which assumes that the buffer names does not change and is always
equal to eir-shell-buffer-name. Specifically, each call to
eir-eval-in-shell causes a new buffer to be created.

This change prevents the dynamic name changes for vterm buffers
created by eval-in-repl.